refactor: 多后端完全支撑——JS 语义迁入 builtins、HeapAccessV2 下沉 wjsm-gc、JsBackend trait 接线
P1-P8: ECMAScript/WHATWG 语义算法全量迁入 wjsm-builtins（<E: ExecContext> 泛型单态化），
host_imports 缩为薄注册层。涵盖 promise combinator、render、json、core、Atomics、
Streams 全家、Fetch、Modules/CJS、collections_buffers、gc property 路径、array_object。
runtime_json + runtime_render 迁移；RegExp 胶水与 Node I/O 桥留 host-wasm。

P9: HeapAccessV2 泛型化迁入 wjsm-gc（HeapAccessV2<M: GrowableHeapMemory>），
host-wasm 侧类型别名保持 callsite 零改动。HeapMemory trait 增加 read_c_string 默认实现。

P10: wjsm-host 新增 JsBackend trait（compile/artifact_bytes/execute，静态分发），
WasmBackend（host-wasm）+ JitBackend（stub）实现，CLI Target::Jit 不再 bail。

P11: exec_context_impl.rs 5491 行拆分为 17 子模块（macro_rules + include 模式），
重写 backend-implementation-guide.md 为六步接入手册，新增 ADR 0013。

全部 1805 测试通过；GC 矩阵 mark-sweep/g1/zgc 各 667 happy 通过；耦合断言全 0。

+++ crates/wjsm-builtins/src/typedarray_methods.rs
@@ -271,7 +271,7 @@ pub fn typedarray_proto_length<E: ExecContext>(ctx: &mut E, this_val: Value) -> 
 
 pub fn typedarray_proto_byte_length<E: ExecContext>(ctx: &mut E, this_val: Value) -> Value {
     ctx.typedarray_resolve(this_val)
-        .map(|v| value::encode_f64((v.length as u32 * v.element_size as u32) as f64))
+        .map(|v| value::encode_f64((v.length * v.element_size as u32) as f64))
         .unwrap_or_else(|| value::encode_f64(0.0))
 }
